+++ EngineSource/Networking/Packet.cpp
@@ -69,15 +69,8 @@ void Packet::EvaluatePacket()
 			}
 			Value.append({ (char)Data[i] });
 		}
-		std::vector<std::string> Values;
-		size_t Last = 0;
 
-		do
-		{
-			size_t PrevLast = Last == 0 ? Last : Last + 1;
-			Last = Value.find_first_of(";", PrevLast);
-			Values.push_back(Value.substr(PrevLast, Last - PrevLast));
-		} while (Last != std::string::npos);
+		auto Values = StrUtil::SeperateString(Value, ';');
 
 		for (auto& i : Values)
 		{