+++ EngineSource/Objects/CSharpObject.cpp
@@ -1,6 +1,15 @@
 #ifdef ENGINE_CSHARP
 #include "CSharpObject.h"
 #include <Engine/Log.h>
+#include <Engine/Utility/StringUtility.h>
+#include <map>
+
+static std::map<std::string, Type::TypeEnum> ManagedTypes =
+{
+	{"String", Type::String},
+	{"Float", Type::Float},
+	{"Vector3", Type::Vector3}
+};
 
 void CSharpObject::Begin()
 {
@@ -25,8 +34,13 @@ void CSharpObject::Destroy()
 	CSharp::DestroyObject(CS_Obj);
 }
 
-void CSharpObject::Reload()
+void CSharpObject::Reload(bool DeleteParameters)
 {
+	if (DeleteParameters)
+	{
+		Properties.clear();
+		AddEditorProperty(Property("C#:Object class", Type::String, &CSharpClass));
+	}
 	OldCSharpClass = CSharpClass;
 	if (CS_Obj.ID)
 	{
@@ -35,7 +49,44 @@ void CSharpObject::Reload()
 	CS_Obj = CSharp::InstantiateObject(CSharpClass, GetTransform(), this);
 	if (CS_Obj.ID)
 	{
+		auto LoadedProperties = StrUtil::SeperateString(CSharp::ExectuteStringFunctionOnObject(CS_Obj, "GetEditorProperties"), ';');
+
+		if (!DeleteParameters)
+		{
+			for (auto& i : Properties)
+			{
+				if (i.PType == Property::PropertyType::CSharpProperty)
+				{
+					static_cast<CSharpObject*>(this)->SetProperty(i.Name.substr(i.Name.find_last_of(":") + 1), i.ValueString);
+				}
+			}
+			Properties.clear();
+			AddEditorProperty(Property("C#:Object class", Type::String, &CSharpClass));
+		}
 		CSharp::ExectuteFunctionOnObject(CS_Obj, "Begin");
+		for (auto& i : LoadedProperties)
+		{
+			size_t FirstSpace = i.find_first_of(" ");
+			std::vector<std::string> values = {i.substr(0, FirstSpace), i.substr(FirstSpace + 1)};
+			bool Array = values[0].size() > 2 && values[0].substr(values[0].size() - 2) == "[]";
+			if (Array)
+			{
+				values[0] = values[0].substr(0, values[0].size() - 2);
+			}
+
+			if (!ManagedTypes.contains(values[0]))
+			{
+				Log::Print("Unknown managed type: " + values[0], Log::LogColor::Red);
+				continue;
+			}
+			Property p = Property(values[1], ManagedTypes[values[0]], nullptr);
+			if (Array)
+			{
+				p.Type = (Type::TypeEnum)(p.Type | Type::List);
+			}
+			p.PType = Property::PropertyType::CSharpProperty;
+			Properties.push_back(p);
+		}
 		if (Name == "CSharpObject")
 		{
 			Name = CSharpClass;
@@ -47,18 +98,28 @@ void CSharpObject::Reload()
 	}
 }
 
+std::string CSharpObject::GetProperty(std::string PropertyName)
+{
+	return CSharp::GetPropertyOfObject(CS_Obj, PropertyName);
+}
+
+void CSharpObject::SetProperty(std::string PropertyName, std::string Value)
+{
+	return CSharp::SetPropertyOfObject(CS_Obj, PropertyName, Value);
+}
+
 void CSharpObject::OnPropertySet()
 {
 	if (OldCSharpClass != CSharpClass)
 	{
-		Reload();
+		Reload(Properties.empty() || !OldCSharpClass.empty());
 	}
 }
 
 void CSharpObject::LoadClass(std::string ClassName)
 {
 	CSharpClass = ClassName;
-	Reload();
+	Reload(true);
 }
 
 #endif