+++ EngineSource/UI/EditorUI/ContextMenu.cpp
@@ -14,6 +14,7 @@
 #include <Engine/Utility/StringUtility.h>
 #include <Engine/Application.h>
 #include "ObjectList.h"
+#include <Objects/CSharpObject.h>
 
 ContextMenu::ContextMenu(EditorPanel* Parent, bool IsScene) : EditorPanel(Parent, IsScene ? "Scene" : "Object Properties")
 {
@@ -82,6 +83,84 @@ std::vector<T>& GetVec(void* vec)
 	return *(std::vector<T>*)vec;
 }
 
+void ContextMenu::GenerateCSharpProperty(const ContextMenu::ContextMenuSection& Element, WorldObject* ContextObject)
+{
+	UIBox* NewElement = nullptr;
+	CSharpObject* obj = static_cast<CSharpObject*>(ContextObject);
+	std::string Value = obj->GetProperty(Element.Name);
+	UIVectorField::VecType VectorType = UIVectorField::VecType::xyz;
+
+	if (Value.find("\r") == std::string::npos)
+	{
+		switch (Element.Type)
+		{
+		case Type::Vector3Color:
+			VectorType = UIVectorField::VecType::rgb;
+			[[fallthrough]];
+		case Type::Vector3Rotation:
+			if (VectorType == UIVectorField::VecType::xyz)
+			{
+				VectorType = UIVectorField::VecType::PitchYawRoll;
+			}
+			[[fallthrough]];
+		case Type::Vector3:
+			NewElement = new UIVectorField(Scale.X - 0.04f, Vector3::FromString(Value), this, -1, EditorUI::Text);
+			NewElement->SetPadding(0.005f, 0, 0.02f, 0);
+			((UIVectorField*)NewElement)->SetValueType(VectorType);
+			break;
+		case Type::String:
+		case Type::Float:
+		case Type::Int:
+			NewElement = GenerateTextField(Value, -1);
+			break;
+		default:
+			break;
+		}
+		if (NewElement)
+		{
+			BackgroundBox->AddChild(NewElement);
+			ContextButtons.push_back(NewElement);
+			ContextSettings.push_back(Element);
+		}
+		return;
+	}
+	return;
+	auto Elements = StrUtil::SeperateString(Value, '\r');
+	for (auto& i : Elements)
+	{
+		switch (Element.Type & ~Type::List)
+		{
+		case Type::Vector3Color:
+			VectorType = UIVectorField::VecType::rgb;
+			[[fallthrough]];
+		case Type::Vector3Rotation:
+			if (VectorType == UIVectorField::VecType::xyz)
+			{
+				VectorType = UIVectorField::VecType::PitchYawRoll;
+			}
+			[[fallthrough]];
+		case Type::Vector3:
+			NewElement = new UIVectorField(Scale.X - 0.04f, Vector3::FromString(i), this, -1, EditorUI::Text);
+			NewElement->SetPadding(0.005f, 0, 0.02f, 0);
+			((UIVectorField*)NewElement)->SetValueType(VectorType);
+			break;
+		case Type::String:
+		case Type::Float:
+		case Type::Int:
+			NewElement = GenerateTextField(i, -1);
+			break;
+		default:
+			break;
+		}
+		if (NewElement)
+		{
+			BackgroundBox->AddChild(NewElement);
+			ContextButtons.push_back(NewElement);
+		}
+	}
+	ContextSettings.push_back(Element);
+}
+
 void ContextMenu::GenerateSectionElement(ContextMenuSection Element, WorldObject* ContextObject, std::string Name)
 {
 	UIBox* NewElement = nullptr;
@@ -95,7 +174,7 @@ void ContextMenu::GenerateSectionElement(ContextMenuSection Element, WorldObject
 
 	size_t NumElements = 1;
 
-	if (IsList)
+	if (IsList && Element.Variable)
 	{
 		switch (Element.Type & ~Type::List)
 		{
@@ -125,6 +204,11 @@ void ContextMenu::GenerateSectionElement(ContextMenuSection Element, WorldObject
 	{
 		void* val = Element.Variable;
 
+		if (val == nullptr)
+		{
+			GenerateCSharpProperty(Element, ContextObject);
+			continue;
+		}
 		if (IsList)
 		{
 			switch (Element.Type & ~Type::List)
@@ -237,6 +321,47 @@ void ContextMenu::OnButtonClicked(int Index)
 
 		for (size_t i = 0; i < ContextButtons.size(); i++)
 		{
+			if (!ContextSettings[IteratedElement].Variable)
+			{
+				auto& Element = ContextSettings[IteratedElement];
+				CSharpObject* obj = static_cast<CSharpObject*>(EditorUI::SelectedObjects[0]);
+
+				if (ContextSettings[IteratedElement].Type & Type::List)
+				{
+					auto arr = StrUtil::SeperateString(obj->GetProperty(Element.Name), '\r');
+					std::string Value;
+					for (auto& arrElem : arr)
+					{
+						ENGINE_ASSERT(arrElem.find("\r") == std::string::npos, "");
+						Value.append(((UITextField*)ContextButtons[i])->GetText());
+						i++;
+					}
+					Value.pop_back();
+					i--;
+					obj->SetProperty(Element.Name, Value);
+				}
+				else
+				{
+					switch (Element.Type)
+					{
+					case Type::Int:
+					case Type::Float:
+					case Type::String:
+						obj->SetProperty(Element.Name, ((UITextField*)ContextButtons[i])->GetText());
+						break;
+					case Type::Vector3:
+					case Type::Vector3Color:
+					case Type::Vector3Rotation:
+						obj->SetProperty(Element.Name, ((UIVectorField*)ContextButtons[i])->GetValue().ToString());
+						break;
+					default:
+						break;
+					}
+				}
+
+				IteratedElement++;
+				continue;
+			}
 			if (ContextSettings[IteratedElement].Type & Type::List)
 			{
 				size_t NumElements = 1;
@@ -305,6 +430,7 @@ void ContextMenu::OnButtonClicked(int Index)
 					}
 					i++;
 				}
+				i--;
 			}
 			else
 			{
@@ -406,7 +532,8 @@ void ContextMenu::OnResized()
 
 		for (WorldObject::Property i : SelectedObject->Properties)
 		{
-			if (i.PType != WorldObject::Property::PropertyType::EditorProperty)
+
+			if (i.PType != WorldObject::Property::PropertyType::EditorProperty && i.PType != WorldObject::Property::PropertyType::CSharpProperty)
 			{
 				continue;
 			}
@@ -430,6 +557,9 @@ void ContextMenu::OnResized()
 				Categories[CategoryName].push_back(ContextMenuSection(i.Data, i.Type, i.Name));
 			}
 		}
+
+		CSharpObject* CSharp = dynamic_cast<CSharpObject*>(SelectedObject);
+
 		char Iterator = 1;
 		for (auto& i : Categories)
 		{